A bunch of changes that fix #40
