+++ lib/lib612/MotionProfile.cpp
