+++ lib/lib612/MotionProfile.h
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace lib612 {
+    class MotionProfile {
+    /*
+    Not to be confused with a drive profile
+    MotionProfile stores target velocity for left and right at a specific timemark
+
+    */
+    public:
+
+    };
+}
