--- lib612/MotionProfile.h
@@ -1,11 +1,0 @@
-#pragma once
-
-class MotionProfile {
-/*
-Not to be confused with a drive profile
-MotionProfile stores target velocity for left and right at a specific timemark
-
-*/
-public:
-
-};
