+++ script/win/.get-ctre.bat
