+++ script/win/deploy.bat
