+++ src/Robot.cpp
@@ -4,7 +4,7 @@
 #include "Commands/Test/SystemCheck.h"
 #include "Commands/Test/TalonTest.h"
 #include "Commands/Autonomous/Autonomous.h"
-#include "../lib612/DriveProfile.h"
+#include "lib612/DriveProfile.h"
 
 std::shared_ptr<Shooter> Robot::shooter;
 std::shared_ptr<Drivetrain> Robot::drivetrain;
@@ -15,8 +15,6 @@ std::unique_ptr<Command> Robot::AutoDrive;
 std::unique_ptr<Command> Robot::drivecommand;
 std::unique_ptr<Command> Robot::CheckSystem;
 std::unique_ptr<Command> Robot::talontesttest;
-
-
 
 void Robot::RobotInit() {
     RobotMap::init();
