+++ src/Subsystems/Drivetrain.h
@@ -3,7 +3,7 @@
 #include "Commands/Subsystem.h"
 #include "WPILib.h"
 #include "CANTalon.h"
-#include "../../lib612/DriveProfile.h"
+#include "lib612/DriveProfile.h"
 #include "../Ports.h"
 
 class Drivetrain: public Subsystem {
@@ -12,13 +12,13 @@ public:
     std::shared_ptr<CANTalon> drive_rl;
     std::shared_ptr<CANTalon> drive_fr;
     std::shared_ptr<CANTalon> drive_rr;
-    DriveProfile* profile;
+    lib612::DriveProfile* profile;
     bool Debug; //if true, debuging is on
     Drivetrain();
-    Drivetrain(DriveProfile* dp);
-    Drivetrain(DriveProfile* dp, bool d);
+    Drivetrain(lib612::DriveProfile* dp);
+    Drivetrain(lib612::DriveProfile* dp, bool d);
     void Init();
-    void SetDriveProfile(DriveProfile dp);
+    void SetDriveProfile(lib612::DriveProfile dp);
     void SetVelocity(double l, double r); //in meters per second (tangentially)
     double GetLeftVelocity();
     double GetRightVelocity();
